# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Broker API",
    "version": "1.0.0",
    "description": "This is the APIs documentation for Broker Platform"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Development server" },
    { "url": "https://broker-six-navy.vercel.app/", "description": "Production server" }
  ],
...
  "paths": {
...
    "/api/uploads/cloudinary": {
      "post": {
        "tags": ["Uploads"],
        "summary": "Upload a file to Cloudinary",
        "description": "Accepts multipart form data and relays the file to Cloudinary using the platform's credentials.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "File to upload. Limited to 10MB by the server."
                  },
                  "field": {
                    "type": "string",
                    "description": "Optional context describing the consumer field (e.g., passportPhoto, idDocument)."
                  }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "File uploaded successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CloudinaryUploadResponse" }
              }
            }
          },
          "400": {
            "description": "No file provided or invalid payload"
          },
          "413": {
            "description": "File size exceeds server limit"
          },
          "500": {
            "description": "Unexpected upload failure"
          }
        }
      }
    },
...
    "/api/forms/PurchaseOrderForm": {
      "get": {
        "tags": ["Order Forms"],
        "summary": "List purchase orders",
        "responses": {
          "200": {
            "description": "Purchase orders retrieved",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PurchaseOrderListResponse" }
              }
            }
          },
          "500": { "description": "Server error" }
        }
      },
      "post": {
        "tags": ["Order Forms"],
        "summary": "Create a purchase order",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PurchaseOrderRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Purchase order created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PurchaseOrderResponse" }
              }
            }
          },
          "400": { "description": "Validation error" },
          "409": { "description": "Conflict - duplicate order" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/api/forms/PurchaseOrderForm/{id}": {
      "get": {
        "tags": ["Order Forms"],
        "summary": "Get purchase order by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Purchase order found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PurchaseOrderResponse" }
              }
            }
          },
          "404": { "description": "Purchase order not found" },
          "500": { "description": "Server error" }
        }
      },
      "patch": {
        "tags": ["Order Forms"],
        "summary": "Update purchase order",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PurchaseOrderUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Purchase order updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PurchaseOrderResponse" }
              }
            }
          },
          "400": { "description": "Validation error" },
          "404": { "description": "Purchase order not found" },
          "500": { "description": "Server error" }
        }
      },
      "delete": {
        "tags": ["Order Forms"],
        "summary": "Delete purchase order",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Purchase order deleted" },
          "404": { "description": "Purchase order not found" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/api/forms/SaleOrderForm": {
      "get": {
        "tags": ["Order Forms"],
        "summary": "List sale orders",
        "responses": {
          "200": {
            "description": "Sale orders retrieved",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SaleOrderListResponse" }
              }
            }
          },
          "500": { "description": "Server error" }
        }
      },
      "post": {
        "tags": ["Order Forms"],
        "summary": "Create a sale order",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SaleOrderRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sale order created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SaleOrderResponse" }
              }
            }
          },
          "400": { "description": "Validation error" },
          "409": { "description": "Conflict - duplicate order" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/api/forms/SaleOrderForm/{id}": {
      "get": {
        "tags": ["Order Forms"],
        "summary": "Get sale order by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Sale order found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SaleOrderResponse" }
              }
            }
          },
          "404": { "description": "Sale order not found" },
          "500": { "description": "Server error" }
        }
      },
      "patch": {
        "tags": ["Order Forms"],
        "summary": "Update sale order",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SaleOrderUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sale order updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SaleOrderResponse" }
              }
            }
          },
          "400": { "description": "Validation error" },
          "404": { "description": "Sale order not found" },
          "500": { "description": "Server error" }
        }
      },
      "delete": {
        "tags": ["Order Forms"],
        "summary": "Delete sale order",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Sale order deleted" },
          "404": { "description": "Sale order not found" },
          "500": { "description": "Server error" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CloudinaryUploadResponse": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "format": "uri",
            "description": "Secure URL for the uploaded asset"
          },
          "publicId": {
            "type": "string",
            "description": "Cloudinary public identifier"
          },
          "bytes": {
            "type": "integer",
            "format": "int64",
            "description": "Size of the uploaded file in bytes"
          },
          "format": {
            "type": "string",
            "description": "File format as determined by Cloudinary"
          },
          "resourceType": {
            "type": "string",
            "description": "Cloudinary resource type (image, raw, video, etc.)"
          },
          "folder": {
            "type": "string",
            "description": "Cloudinary folder that stores the asset"
          }
        },
        "required": ["url", "publicId", "bytes", "format", "resourceType", "folder"]
      },
...
      "PurchaseOrderItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "nullable": true },
          "security": { "type": "string" },
          "quantity": { "type": "integer", "minimum": 1 },
          "price": {
            "type": "string",
            "nullable": true,
            "description": "Stored as a stringified decimal"
          }
        },
        "required": ["security", "quantity"]
      },
      "SaleOrderItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "nullable": true },
          "security": { "type": "string" },
          "quantity": { "type": "integer", "minimum": 1 },
          "price": {
            "type": "string",
            "nullable": true,
            "description": "Stored as a stringified decimal"
          }
        },
        "required": ["security", "quantity"]
      },
      "PurchaseOrderRequest": {
        "type": "object",
        "properties": {
          "clientName": { "type": "string" },
          "csdNumber": { "type": "string", "nullable": true },
          "phone": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "address": { "type": "string" },
          "bestMarketPrice": { "type": "boolean" },
          "priceLimit": { "type": "boolean" },
          "standingOrderNote": { "type": "string", "nullable": true },
          "standingFrequency": {
            "type": "string",
            "nullable": true,
            "description": "Standing order frequency code"
          },
          "additionalInstructions": { "type": "string", "nullable": true },
          "termsAccepted": { "type": "boolean" },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "security": { "type": "string" },
                "quantity": { "type": "integer", "minimum": 1 },
                "price": { "type": "string", "nullable": true }
              },
              "required": ["security", "quantity"]
            },
            "minItems": 1
          }
        },
        "required": [
          "clientName",
          "phone",
          "email",
          "address",
          "termsAccepted",
          "items"
        ]
      },
      "PurchaseOrderUpdateRequest": {
        "type": "object",
        "minProperties": 1,
        "additionalProperties": false,
        "properties": {
          "clientName": { "type": "string" },
          "csdNumber": { "type": "string", "nullable": true },
          "phone": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "address": { "type": "string" },
          "bestMarketPrice": { "type": "boolean" },
          "priceLimit": { "type": "boolean" },
          "standingOrderNote": { "type": "string", "nullable": true },
          "standingFrequency": {
            "type": "string",
            "nullable": true,
            "description": "Standing order frequency code"
          },
          "additionalInstructions": { "type": "string", "nullable": true },
          "termsAccepted": { "type": "boolean" },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PurchaseOrderItem" }
          }
        }
      },
      "SaleOrderRequest": {
        "type": "object",
        "properties": {
          "clientName": { "type": "string" },
          "csdNumber": { "type": "string", "nullable": true },
          "phone": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "address": { "type": "string" },
          "bestMarketPrice": { "type": "boolean" },
          "priceLimit": { "type": "boolean" },
          "withinTimeLimitNote": { "type": "string", "nullable": true },
          "bankName": { "type": "string", "nullable": true },
          "bankBranch": { "type": "string", "nullable": true },
          "accountNumber": { "type": "string", "nullable": true },
          "termsAccepted": { "type": "boolean" },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "security": { "type": "string" },
                "quantity": { "type": "integer", "minimum": 1 },
                "price": { "type": "string", "nullable": true }
              },
              "required": ["security", "quantity"]
            },
            "minItems": 1
          }
        },
        "required": [
          "clientName",
          "phone",
          "email",
          "address",
          "termsAccepted",
          "items"
        ]
      },
      "SaleOrderUpdateRequest": {
        "type": "object",
        "minProperties": 1,
        "additionalProperties": false,
        "properties": {
          "clientName": { "type": "string" },
          "csdNumber": { "type": "string", "nullable": true },
          "phone": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "address": { "type": "string" },
          "bestMarketPrice": { "type": "boolean" },
          "priceLimit": { "type": "boolean" },
          "withinTimeLimitNote": { "type": "string", "nullable": true },
          "bankName": { "type": "string", "nullable": true },
          "bankBranch": { "type": "string", "nullable": true },
          "accountNumber": { "type": "string", "nullable": true },
          "termsAccepted": { "type": "boolean" },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SaleOrderItem" }
          }
        }
      },
      "PurchaseOrder": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "clientName": { "type": "string" },
          "csdNumber": { "type": "string", "nullable": true },
          "phone": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "address": { "type": "string" },
          "bestMarketPrice": { "type": "boolean" },
          "priceLimit": { "type": "boolean" },
          "standingOrderNote": { "type": "string", "nullable": true },
          "standingFrequency": { "type": "string", "nullable": true },
          "additionalInstructions": { "type": "string", "nullable": true },
          "termsAccepted": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PurchaseOrderItem" }
          }
        }
      },
      "SaleOrder": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "clientName": { "type": "string" },
          "csdNumber": { "type": "string", "nullable": true },
          "phone": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "address": { "type": "string" },
          "bestMarketPrice": { "type": "boolean" },
          "priceLimit": { "type": "boolean" },
          "withinTimeLimitNote": { "type": "string", "nullable": true },
          "bankName": { "type": "string", "nullable": true },
          "bankBranch": { "type": "string", "nullable": true },
          "accountNumber": { "type": "string", "nullable": true },
          "termsAccepted": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SaleOrderItem" }
          }
        }
      },
      "PurchaseOrderResponse": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/PurchaseOrder" }
        },
        "required": ["data"]
      },
      "SaleOrderResponse": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/SaleOrder" }
        },
        "required": ["data"]
      },
      "PurchaseOrderListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PurchaseOrder" }
          }
        },
        "required": ["data"]
      },
      "SaleOrderListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SaleOrder" }
          }
        },
        "required": ["data"]
      }
    }
  }
}